README: how to test the application moved antifraud  app.module to application directory (clean architecture)
