+++ antifraud_project/src/application/app.module.ts
@@ -2,13 +2,13 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { MessageConsumerController } from './adapter/input/messaging/message_consumer.controller';
-import { TransactionEntity } from './domain/models/transaction.entity';
-import { FraudAnalysisUsecase } from './domain/usecases/fraud_analysis.usecase';
-import { TransactionService } from './adapter/output/db/transaction.service';
-import { MessagingService } from './adapter/input_output/messaging/messaging.service';
-import { MessageProducerInitializer } from './adapter/output/messaging/message_producer.initializer';
-import configurationYaml from '../configuration.yaml';
+import { MessageConsumerController } from '../adapter/input/messaging/message_consumer.controller';
+import { TransactionEntity } from '../domain/models/transaction.entity';
+import { FraudAnalysisUsecase } from '../domain/usecases/fraud_analysis.usecase';
+import { TransactionService } from '../adapter/output/db/transaction.service';
+import { MessagingService } from '../adapter/input_output/messaging/messaging.service';
+import { MessageProducerInitializer } from '../adapter/output/messaging/message_producer.initializer';
+import configurationYaml from '../../configuration.yaml';
 @Module({
   imports: [
     ConfigModule.forRoot({ load: [configurationYaml] }),
